requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.2.0
click==8.1.7
discord.py==2.3.2
frozenlist==1.4.0
idna==3.4
mihomo @ git+https://github.com/KT-Yeh/mihomo.git@6c6b02cff304d726a9b6acb863f9776c794d6cf2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==1.10.12
typing_extensions==4.8.0
yarl==1.9.2
